pyproject.toml
[tool.poetry]
name = "llm-with-rag-chatbot"
version = "0.1.0"
description = ""
authors = ["Farid Seifi <farid.seifi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-session = "^0.8.0"
openai = "^1.35.14"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
docarray = "^0.40.0"
flask-caching = "^2.3.0"
evaluate = "^0.4.3"
nltk = "^3.9.1"
rouge-score = "^0.1.2"
qdrant-client = "^1.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
